Address project registry review feedback
